[REF] Refactor segmentation code into many files and pimms system
