+++ AFQ/utils/bin.py
@@ -165,15 +165,16 @@ def dict_to_json(dictionary):
 
 def func_dict_to_arg_dict(func_dict=None, logger=None):
     if func_dict is None:
-        import AFQ.segmentation as seg
+        from AFQ.bundle_rec.recognize import recognize
+        from AFQ.bundle_rec.cleaning import clean_bundle
         import AFQ.tractography.tractography as aft
         from AFQ.api.group import GroupAFQ
 
         func_dict = {
             "BIDS": GroupAFQ.__init__,
             "Tractography": aft.track,
-            "Segmentation": seg.Segmentation.__init__,
-            "Cleaning": seg.clean_bundle}
+            "Segmentation": recognize,
+            "Cleaning": clean_bundle}
 
     arg_dict = {}
     for name, func in func_dict.items():
@@ -411,13 +412,14 @@ def generate_json(json_folder, overwrite=False,
         }
     ]
 }"""
-    import AFQ.segmentation as seg
+    from AFQ.bundle_rec.recognize import recognize
+    from AFQ.bundle_rec.cleaning import clean_bundle
     import AFQ.tractography.tractography as aft
 
     func_dict = {
         "Tractography": aft.track,
-        "Segmentation": seg.Segmentation.__init__,
-        "Cleaning": seg.clean_bundle}
+        "Segmentation": recognize,
+        "Cleaning": clean_bundle}
 
     arg_dict = func_dict_to_arg_dict(func_dict, logger=logger)
